The server now takes the site on the first line of the communication and selects the right config using it. Updated all the writes to the server to include the site.
The init script now uses the site name as the database name.

Made some interface changes: after a new user is created the user is taken back
to the admin home page, which now displays all users. After a friend request
the user is taken back to the friend's home page.

Made a fix to the makefile clean target.