Compile/Link against openssl in /opt/openssl if it is there. Using a from-source version allows removal of valgrind false-positives.
Removed the printfs in disttree.cpp (actually replaced with debug stub). If the
stdout buffer filled and this got flushed it would pollute the SSL stream out.
Believe to be the cause of the symptoms that motivated ticket #26.

The ftr command changed to forward_tree_reset.

Fixed some mismatched new/delete/malloc/free.