There are now three types of notifications.
The user_message notification is for direct messages that involve just a sender
and receiver.

(new) The user_message_double is for user messages that that involve remote
activity (has subject and author). Probably should rename this.

The remote_broadcast is for notification of remote activities sent to the
author.