Added a script that generates the system install instructions. It ask what the webserver user is.
There is now a config file for the new-site script. The www-user goes in it, as
well as the CA cert file.

The dsnpd.conf file step is now separated into two steps. One to install the
cert and key, the other to install the config fragment. Since using the new
site config file, nothing needs to change in the dsnpd.conf file fragment. It
can go in as is.